Move NIST SP-1500-102 models into 'models.nist.vri' sub-package. Update all tests to match.
