+++ test/vanadium/tests/models/test_vri_models.py
@@ -6,7 +6,7 @@ from contextlib import nullcontext as raises_none
 import pytest
 from pydantic import ValidationError
 
-from vanadium.models import *
+from vanadium.models.nist.vri import *
 
 from tests.conftest import Raises
